requirements.txt
requests
python-dotenv
pyyaml
pydantic>=2.0.0
openai>=1.0.0
